+++ src/error.rs
@@ -153,6 +153,7 @@ win32_errors! {
     ERROR_PIPE_BUSY = 231 => "all pipe instances are busy",
     ERROR_PIPE_CONNECTED = 535 => "a client is connected to the pipe",
     ERROR_NOTIFY_ENUM_DIR = 1022 => "the buffer for the directory-change notifications overflowed",
+    ERROR_NO_MORE_ITEMS = 259 => "no more data is available",
 }
 
 impl core::error::Error for Win32Error {}